query-engine/worker/requirements.txt
grpcio
grpcio-tools
psycopg2-binary
sqlparse